requirements.txt
pillow==10.2.0
matplotlib==3.8.2
timm==0.9.12
torch==2.2.0
transformers==4.37.2
python-dotenv==1.0.1
open-clip-torch==2.24.0
marshmallow-dataclass==8.5.14
omegaconf==2.3.0
hydra-core==1.3.2
typeguard
qdrant-client==1.7.1
streamlit==1.30.0